data_processing/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Database connectivity
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Data validation and type checking
pydantic>=2.0.0

# Progress bars and logging
tqdm>=4.65.0
loguru>=0.7.0

# Statistical analysis
scipy>=1.10.0
scikit-learn>=1.3.0

# Date/time handling
python-dateutil>=2.8.0

# Environment variables
python-dotenv>=1.0.0